Update to static plugin
An update to the static file server plugin that addresses issues found in

https://github.com/mcavage/node-restify/issues/420

as well as updates to the documentation and tests related to the issue.

This updates the behavior for static file serving to use more intuitive
directory linking where /some/path/ can easily be mapped to /another/place/
including all subdirectories.